requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
backports.entry-points-selectable==1.1.0
blinker==1.4
Brotli==1.0.9
case==1.5.3
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
colorlog==6.5.0
cryptography==3.4.8
distlib==0.3.2
et-xmlfile==1.1.0
Faker==10.0.0
filelock==3.0.12
Flask==2.0.2
h11==0.12.0
h2==4.1.0
har2case==0.3.1
hpack==4.0.0
HttpRunner==1.5.8
hyperframe==6.0.1
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonpath==0.82
kaitaistruct==0.9
ldap3==2.9.1
linecache2==1.0.0
MarkupSafe==2.0.1
mitmproxy==7.0.4
mock==4.0.3
msgpack==1.0.3
nose==1.3.7
openpyxl==3.0.9
packaging==21.3
passlib==1.7.4
pika==1.2.0
pipenv==2021.5.29
platformdirs==2.2.0
pluggy==1.0.0
protobuf==3.18.1
publicsuffix2==2.20191221
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydivert==2.1.0
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
pytest==6.2.5
pytest-mock==3.6.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.26.0
requests-toolbelt==0.9.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sortedcontainers==2.4.0
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traceback2==1.4.0
unittest2==1.1.0
urllib3==1.26.7
urwid==2.1.2
virtualenv==20.7.0
virtualenv-clone==0.5.6
Werkzeug==2.0.2
wsproto==1.0.0
xlwt==1.3.0
zstandard==0.15.2
